Update build tools

+++ publish.sh
@@ -1,8 +1,13 @@
 #!/usr/bin/env bash
 
 set -e
+set -u
 
-TIMESTAMP=$(date +"%Y-%m-%dT%H:%M:%S")
+timestamp="$(date +"%Y-%m-%dT%H:%M:%S")"
+src_branch="$(git branch | cut -d ' ' -f2)"
+
+echo "Warning: All submodules will be updated, any local changes may be overwritten!"
+read -p "Press ENTER to continue, or CTRL+C to cancel"
 
 # Make sure submodules are updated
 echo "Updating submodules ..."
@@ -17,14 +22,20 @@ cd ..
 # Build
 echo "Building site ..."
 hugo -v
-echo "Build successful! (But do check for errors above manually.)"
+echo "Build successful! (But do check for errors above manually)"
 
 # Publish to master
 echo
 echo "Do you wish to publish the built site?"
-read -p "Press ENTER to continue, or CTRL+C to cancel."
+read -p "Press ENTER to continue, or CTRL+C to cancel"
+
 cd public
+publish_branch="$(git branch | cut -d ' ' -f2)"
+if [[ $src_branch == $publish_branch ]]; then
+    echo "Error: Source and publish use the same branch" 1>&2
+    exit -1
+fi
 git add -A
-git commit --allow-empty -m "Build $TIMESTAMP"
+git commit --allow-empty -m "Build $timestamp"
 git push --force
 echo "Publish successful!"
